+++ src/main/scala/co/topl/brambl/cli/BramblCliParamsParserModule.scala
@@ -246,15 +246,41 @@ object BramblCliParamsParserModule {
       opt[Option[Int]]("from-state")
         .action((x, c) => c.copy(someFromState = x))
         .text("State from where we are sending the funds from"),
+      opt[Option[String]]("change-party")
+        .action((x, c) => c.copy(someChangeParty = x))
+        .text("Party where we are sending the change to"),
+      opt[Option[String]]("change-contract")
+        .action((x, c) => c.copy(someChangeContract = x))
+        .text("Contract where we are sending the change to"),
+      opt[Option[Int]]("change-state")
+        .action((x, c) => c.copy(someChangeState = x))
+        .text("State where we are sending the change to"),
       checkConfig(c =>
         if (c.fromParty == "noparty") {
           if (c.someFromState.isEmpty) {
             failure("You must specify a from-state when using noparty")
           } else {
+            (c.someChangeParty, c.someChangeContract, c.someChangeState) match {
+              case (Some(_), Some(_), Some(_)) =>
+                success
+              case (_, _, _) =>
+                failure(
+                  "You must specify a change-party, change-contract and change-state when using noparty"
+                )
+            }
             success
           }
         } else {
-          success
+          (c.someChangeParty, c.someChangeContract, c.someChangeState) match {
+            case (Some(_), Some(_), Some(_)) =>
+              success
+            case (None, None, None) =>
+              success
+            case (_, _, _) =>
+              failure(
+                "You must specify a change-party, change-contract and change-state or not specify any of them"
+              )
+          }
         }
       )
     )
@@ -696,7 +722,7 @@ object BramblCliParamsParserModule {
       tokenType: TokenType.Value,
       groupId: Option[GroupId],
       seriesId: Option[SeriesId]
-  ) = {
+  ): Either[String, Unit] = {
     (tokenType, groupId, seriesId) match {
       case (TokenType.group, Some(_), None) =>
         success