+++ src/main/scala/co/topl/brambl/cli/impl/SimpleTransactionAlgebra.scala
@@ -25,6 +25,9 @@ trait SimpleTransactionAlgebra[F[_]] {
       fromParty: String,
       fromContract: String,
       someFromState: Option[Int],
+      someChangeParty: Option[String],
+      someChangeContract: Option[String],
+      someChangeState: Option[Int],
       someToAddress: Option[LockAddress],
       someToParty: Option[String],
       someToContract: Option[String],
@@ -49,6 +52,9 @@ object SimpleTransactionAlgebra {
 
       private def buildTransaction(
           txos: Seq[Txo],
+          someChangeParty: Option[String],
+          someChangeContract: Option[String],
+          someChangeState: Option[Int],
           predicateFundsToUnlock: Lock.Predicate,
           lockForChange: Lock,
           recipientLockAddress: LockAddress,
@@ -75,8 +81,23 @@ object SimpleTransactionAlgebra {
             )
           ioTransaction <- Sync[F].fromEither(eitherIoTransaction)
           // Only save to wallet state if there is a change output in the transaction
+          nextIndicesExist <- (
+            someChangeParty,
+            someChangeContract,
+            someChangeState
+          ) match {
+            case (Some(changeParty), Some(changeContract), Some(changeState)) =>
+              walletStateApi.getCurrentIndicesForFunds(
+                changeParty,
+                changeContract,
+                Some(changeState)
+              ).map(_.isDefined)
+            case _ => Sync[F].point(false)
+          }
           _ <-
-            if (ioTransaction.outputs.length >= 2) for {
+            if (
+              ioTransaction.outputs.length >= 2 && !nextIndicesExist
+            ) for {
               lockAddress <-
                 transactionBuilderApi.lockAddress(
                   lockForChange
@@ -88,6 +109,7 @@ object SimpleTransactionAlgebra {
                     .map(_.vk)
                 )
                 .sequence
+
               _ <- walletStateApi.updateWalletState(
                 Encoding.encodeToBase58Check(
                   lockForChange.getPredicate.toByteArray
@@ -126,6 +148,9 @@ object SimpleTransactionAlgebra {
           fromParty: String,
           fromContract: String,
           someFromState: Option[Int],
+          someChangeParty: Option[String],
+          someChangeContract: Option[String],
+          someChangeState: Option[Int],
           someToAddress: Option[LockAddress],
           someToParty: Option[String],
           someToContract: Option[String],
@@ -153,21 +178,26 @@ object SimpleTransactionAlgebra {
             )
             .sequence
             .map(_.flatten.map(Lock().withPredicate(_)))
-          someNextIndices <- walletStateApi.getNextIndicesForFunds(
-            if (fromParty == "noparty") "self" else fromParty,
-            if (fromParty == "noparty") "default"
-            else fromContract
-          )
+          someNextIndices <-
+            (someChangeParty, someChangeContract, someChangeState) match {
+              case (Some(party), Some(contract), Some(state)) =>
+                walletStateApi.getCurrentIndicesForFunds(
+                  party,
+                  contract,
+                  Some(state)
+                )
+              case _ =>
+                walletStateApi.getNextIndicesForFunds(
+                  fromParty,
+                  fromContract
+                )
+            }
           // Generate a new lock for the change, if possible
           changeLock <- someNextIndices
-            .map(idx =>
-              walletStateApi.getLock(
-                if (fromParty == "noparty") "self" else fromParty,
-                if (fromParty == "noparty") "default"
-                else fromContract,
-                idx.z
-              )
-            )
+            .map{idx =>
+              walletStateApi
+                .getLock(fromParty, fromContract, idx.z)
+            }
             .sequence
             .map(_.flatten)
           fromAddress <- transactionBuilderApi.lockAddress(
@@ -215,6 +245,9 @@ object SimpleTransactionAlgebra {
                  case (Some(lockPredicateForChange), Some(toAddress)) =>
                    buildTransaction(
                      txos,
+                     someChangeParty,
+                     someChangeContract,
+                     someChangeState,
                      predicateFundsToUnlock.get.getPredicate,
                      lockPredicateForChange,
                      toAddress,