+++ src/main/scala/co/topl/brambl/cli/modules/ContractModeModule.scala
@@ -14,6 +14,8 @@ trait ContractModeModule extends WalletStateResource {
       walletResource(validateParams.walletFile)
     )
     validateParams.subcmd match {
+      case BramblCliSubCmd.invalid =>
+        IO.pure(Left("A subcommand needs to be specified"))
       case BramblCliSubCmd.list =>
         new ContractsController(
           contractStorageAlgebra