+++ src/main/scala/co/topl/brambl/cli/modules/TxModeModule.scala
@@ -12,6 +12,8 @@ trait TxModeModule extends TxParserAlgebraModule with TransactionAlgebraModule {
       validateParams: BramblCliParams
   ): IO[Either[String, String]] = {
     validateParams.subcmd match {
+      case BramblCliSubCmd.invalid =>
+        IO.pure(Left("A subcommand needs to be specified"))
       case BramblCliSubCmd.broadcast =>
         new TxController(
           txParserAlgebra(